requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
sqlalchemy==2.0.28
plotly==5.20.0
streamlit-option-menu==0.3.12
bcrypt==4.1.2
fpdf2==2.7.9
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
